Implement smarter, equal-step panning. Change config format.
Commonly in document viewers, the panning amount is fixed
by the user, so the last pan on a page is clipped to the
page boundary. This makes panning asymmetric (after
trying to go past the page boundary, panning back will
still go a full step) and distributes text suboptimally
between panning steps.

"Equal-step panning" means that we plan ahead for the
number of panning steps per page and distribute
window positions evenly across the page. See source code
for details.

The "pan amount" configuration options are therefore
replaced by a "minimum overlap" option. Further experiments
will show if 10% is a good value for it (5% may be better).